requirements.txt
asgiref==3.5.0
Django==4.0.2
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
pipdeptree==2.2.1
psycopg2==2.9.3
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
